// src/generated-api/models/ProjectCategorySummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BMI-LAB Web API
 * BMI-LAB을 관리하기 위한 웹 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProjectCategorySummary
 */
export interface ProjectCategorySummary {
    /**
     * 카테고리 ID
     * @type {number}
     * @memberof ProjectCategorySummary
     */
    categoryId?: number;
    /**
     * 카테고리 이름
     * @type {string}
     * @memberof ProjectCategorySummary
     */
    name?: string;
}

/**
 * Check if a given object implements the ProjectCategorySummary interface.
 */
export function instanceOfProjectCategorySummary(value: object): value is ProjectCategorySummary {
    return true;
}

export function ProjectCategorySummaryFromJSON(json: any): ProjectCategorySummary {
    return ProjectCategorySummaryFromJSONTyped(json, false);
}

export function ProjectCategorySummaryFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectCategorySummary {
    if (json == null) {
        return json;
    }
    return {
        
        'categoryId': json['categoryId'] == null ? undefined : json['categoryId'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function ProjectCategorySummaryToJSON(json: any): ProjectCategorySummary {
    return ProjectCategorySummaryToJSONTyped(json, false);
}

export function ProjectCategorySummaryToJSONTyped(value?: ProjectCategorySummary | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'categoryId': value['categoryId'],
        'name': value['name'],
    };
}